Overhaul for molotov
This redoes the test harness to use molotov and asyncio.

This picks a commit in molotov master branch between the v0.3 and v0.4 releases
that works pretty well and has a fix to request printing that we needed because
some of the test scenarios send compressed payloads that aren't utf-8 encoded.
When v0.4 comes out, we should switch to that.